@@ -1,6 +1,6 @@
 {
   "name": "notifier-ui",
-  "version": "0.0.0",
+  "version": "2.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -10,35 +10,39 @@
     "e2e": "ng e2e",
     "build-styles": "scss-bundle -e ./projects/addapptables/notifier/src/lib/_addapptable-notifier.theme.scss -d ./dist/addapptables/notifier/_addapptables-notifier.theme.scss"
   },
-  "private": true,
+  "license": "MIT",
+  "homepage": "https://github.com/addapptables/notifier-ui",
+  "private": false,
+  "author": "Addapptables",
+  "description": "Material notifier library for angular",
   "dependencies": {
-    "@addapptables/core": "0.0.1",
-    "@angular/animations": "^7.2.3",
-    "@angular/cdk": "^7.3.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@addapptables/core": "^2.0.0",
+    "@angular/animations": "^8.2.3",
+    "@angular/cdk": "^8.1.3",
+    "@angular/common": "^8.2.3",
+    "@angular/compiler": "^8.2.3",
+    "@angular/core": "^8.2.3",
+    "@angular/forms": "^8.2.3",
+    "@angular/platform-browser": "^8.2.3",
+    "@angular/platform-browser-dynamic": "^8.2.3",
+    "@angular/router": "^8.2.3",
     "core-js": "^2.5.4",
     "perfect-scrollbar": "^1.4.0",
-    "rxjs": "~6.3.3",
+    "rxjs": "6.5.2",
     "scss-bundle": "^2.4.0",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.0",
-    "@angular-devkit/build-ng-packagr": "~0.12.0",
-    "@angular/cli": "~7.2.1",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@types/node": "~8.9.4",
+    "@angular-devkit/build-angular": "^0.803.0",
+    "@angular-devkit/build-ng-packagr": "^0.803.0",
+    "@angular/cli": "^8.3.0",
+    "@angular/compiler-cli": "^8.2.3",
+    "@angular/language-service": "^8.2.3",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
+    "@types/node": "~8.9.4",
+    "codelyzer": "5.1.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.1.1",
@@ -46,12 +50,28 @@
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.2.0",
+    "ng-packagr": "5.5.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tsickle": ">=0.34.0",
+    "tsickle": "0.37.0",
     "tslib": "^1.9.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
-  }
+    "typescript": "3.5.3"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/addapptables/notifier-ui.git"
+  },
+  "bugs": {
+    "url": "https://github.com/addapptables/notifier-ui/issues"
+  },
+  "keywords": [
+    "angular",
+    "angular 2",
+    "angular notifier",
+    "notifier",
+    "material",
+    "addapptables-alert-ui",
+    "addapptables"
+  ]
 }
