@@ -1,16 +1,16 @@
 {
   "name": "@addapptables/notifier",
-  "version": "1.0.2",
+  "version": "2.0.0",
   "author": "Addapptables",
   "description": "Material notifier library for angular",
   "license": "MIT",
   "homepage": "https://github.com/addapptables/notifier-ui",
   "peerDependencies": {
-    "@angular/common": ">=7.0.0 < 8.0.0",
-    "@angular/core": ">=7.0.0 < 8.0.0",
-    "@angular/cdk": ">=7.0.0 < 8.0.0",
-    "@angular/animations": ">=7.0.0 < 8.0.0",
-    "@addapptables/core": ">=1.0.0 < 2.0.0"
+    "@angular/common": ">=8.0.0 < 9.0.0",
+    "@angular/core": ">=8.0.0 < 9.0.0",
+    "@angular/cdk": ">=8.0.0 < 9.0.0",
+    "@angular/animations": ">=8.0.0 < 9.0.0",
+    "@addapptables/core": ">=2.0.0 < 3.0.0"
   },
   "repository": {
     "type": "git",
@@ -20,11 +20,10 @@
     "url": "https://github.com/addapptables/notifier-ui/issues"
   },
   "keywords": [
-    "angular",
-    "angular 2",
+    "angular notifier",
     "notifier",
     "material",
-    "addapptables-alert-ui",
+    "addapptables-notifier-ui",
     "addapptables"
   ]
 }
